+++ package.json
@@ -2,7 +2,7 @@
     "name": "axios-logger",
     "version": "1.0.0",
     "description": "Beautify Axios Logging Messages",
-    "main": "src/index.js",
+    "main": "src/index.ts",
     "keywords": [
         "javascript",
         "nodejs",
@@ -25,6 +25,7 @@
     },
     "homepage": "https://github.com/hg-pyun/axios-intercept-logger#readme",
     "devDependencies": {
+        "@types/node": "^12.0.10",
         "axios": "^0.18.0",
         "husky": "^2.4.1",
         "jest": "^24.8.0",
@@ -33,7 +34,8 @@
         "typescript": "^3.5.2"
     },
     "dependencies": {
-        "chalk": "^2.4.1"
+        "chalk": "^2.4.1",
+        "dateformat": "^3.0.3"
     },
     "scripts": {
         "dev": "nodemon --exec babel-node test/server.js",
