+++ src/common/config.ts
@@ -1,0 +1,8 @@
+function setGlobalConfig() {
+
+}
+
+
+export {
+    setGlobalConfig
+}
