+++ src/common/types.ts
@@ -1,0 +1,3 @@
+export interface logConfig {
+
+}
