--- src/index.js
@@ -1,3 +1,0 @@
-module.exports.requestLogger = require('./logger/request');
-module.exports.responseLogger = require('./logger/response');
-module.exports.errorLogger = require('./logger/error');
