+++ src/index.ts
@@ -1,0 +1,6 @@
+import {setGlobalConfig} from './common/config';
+import requestLogger from './logger/request';
+import responseLogger from './logger/response';
+import errorLogger from './logger/error';
+
+export {setGlobalConfig, requestLogger, responseLogger, errorLogger};
